Rename encoder files
