+++ march_hardware/include/march_hardware/IMotionCube.h
@@ -9,8 +9,8 @@
 #include "march_hardware/ethercat/slave.h"
 #include "march_hardware/IMotionCubeState.h"
 #include "march_hardware/IMotionCubeTargetState.h"
-#include "march_hardware/encoder/AbsoluteEncoder.h"
-#include "march_hardware/encoder/IncrementalEncoder.h"
+#include "march_hardware/encoder/absolute_encoder.h"
+#include "march_hardware/encoder/incremental_encoder.h"
 
 #include <memory>
 #include <unordered_map>
