+++ march_hardware/src/encoder/encoder.cpp
@@ -1,5 +1,5 @@
 // Copyright 2019 Project March.
-#include "march_hardware/encoder/Encoder.h"
+#include "march_hardware/encoder/encoder.h"
 #include "march_hardware/ethercat/pdo_types.h"
 #include "march_hardware/error/hardware_exception.h"
