+++ march_hardware/src/encoder/incremental_encoder.cpp
@@ -1,5 +1,5 @@
 // Copyright 2019 Project March.
-#include "march_hardware/encoder/IncrementalEncoder.h"
+#include "march_hardware/encoder/incremental_encoder.h"
 
 namespace march
 {
