+++ march_hardware/test/mocks/MockAbsoluteEncoder.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "march_hardware/encoder/AbsoluteEncoder.h"
+#include "march_hardware/encoder/absolute_encoder.h"
 
 #include <gmock/gmock.h>
