+++ march_hardware_builder/test/absolute_encoder_builder_test.cpp
@@ -8,7 +8,7 @@
 #include <ros/package.h>
 #include <urdf/model.h>
 
-#include <march_hardware/encoder/AbsoluteEncoder.h>
+#include <march_hardware/encoder/absolute_encoder.h>
 
 class TestAbsoluteEncoderBuilder : public ::testing::Test
 {
